Fix up d4os_ui_users_validate_name() to check for valid OpenSim/SL names in one hit, and use the Drupal name validator instead of duplicating it.
Drupal changed their name validation code since it was originally copied into this function.  So calling it instead of duplicating it's source code makes this more future proof.
The length of OpenSim/SL names was off by one.
This does not put the error message in the correct place, but the next commit fixes that.
Returns a bunch of errors if there's a bunch of things wrong.
See the comments in the code for more.

I'm not sure about the eregi() call, what is that for?